Move “settings” state into “form” subtree.
