+++ client/state/form/reducer.js
@@ -1,7 +1,8 @@
 import { SET_FORM_PROPERTY } from './actions';
 import packages from './packages/reducer';
+import settings from './settings/reducer';
 import * as packagesActions from './packages/actions';
-import * as settingsActions from '../settings/actions';
+import * as settingsActions from './settings/actions';
 
 const reducers = {};
 
@@ -27,13 +28,17 @@ export default function form( state = {}, action ) {
 		} );
 	}
 
-	if ( settingsActions.hasOwnProperty( action.type ) ) {
+	if ( settingsActions.hasOwnProperty( action.type ) || ( packagesActions.SAVE_PACKAGE === action.type ) ) {
 		newState.pristine = false;
 
 		// Allow client-side form validation to take over error state when inputs change
 		if ( newState.hasOwnProperty( 'errors' ) ) {
 			delete newState.errors;
 		}
+
+		newState = Object.assign( newState, {
+			settings: settings( state.settings || {}, action ),
+		} );
 	}
 
 	return newState;
